Refactoring: all non-stress specific code moved to global/root cmd
